refactor/#6 [refactor] : 미작업인 스켈레톤 클래스 리팩토링 작업
