+++ app/src/main/java/com/devnuts/ruflu/ui/mypage/viewmodel/MyPageViewModel.kt
@@ -3,5 +3,5 @@ package com.devnuts.ruflu.ui.mypage.viewmodel
 import androidx.lifecycle.ViewModel
 
 class MyPageViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
+
 }
