+++ app/src/main/res/layout/fragment_account_restore_auth.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+
         <variable
             name="viewModel"
             type="com.devnuts.ruflu.ui.signin.viewmodel.AccountRestoreViewModel" />
@@ -64,8 +65,8 @@
                 android:layout_height="wrap_content"
                 android:text="이메일을 작성해주세요."
                 android:textSize="20sp"
-                app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
 
             <EditText
@@ -73,10 +74,11 @@
                 android:layout_width="match_parent"
                 android:layout_height="50dp"
                 android:layout_marginTop="20dp"
-                android:backgroundTint="@color/black"
+                android:backgroundTint="@color/ruflu_main_orange"
                 android:hint="이메일"
-                android:paddingStart="10dp"
                 android:inputType="textEmailAddress"
+                android:maxLines="1"
+                android:paddingStart="10dp"
                 android:textColor="#000000"
                 android:textColorHint="#A0A0A0"
                 android:textSize="15sp"
