+++ app/src/main/res/layout/fragment_sms.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.devnuts.ruflu.ui.signin.fragment.SMSViewModel" />
+            type="com.devnuts.ruflu.ui.signin.viewmodel.SMSViewModel" />
     </data>
